backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
sqlalchemy==2.0.36
transformers==4.46.3
torch==2.9.1
python-dotenv==1.0.0
google-generativeai==0.3.2
requests>=2.31.0
psycopg[binary]==3.3.2
